# server/src/modules/product/product.controller.js
import { Router } from "express";
import services from "./product.service.js";
import util from "#/common/util/index.js";
import middlewares from "#/common/middleware/index.js";
import DTO from "./dto/index.js";
const router = Router();

router.post("/create", middlewares.validate(DTO.createProductDTO), async (req, res, next) => {
  const serviceFeedback = await services.createProductService(req.body);
  return util.sendSuccess(res, serviceFeedback);
});

router.get("/all", async (req, res, next) => {
  const serviceFeedback = await services.retrieveAllProductService();
  return util.sendSuccess(res, serviceFeedback);
});

router.patch(
  "/update/:productID",
  middlewares.validate(DTO.updateProductDTO),
  async (req, res, next) => {
    const serviceFeedback = await services.updateProductService(req.params.productID, req.body);
    return util.sendSuccess(res, serviceFeedback);
  },
);

router.get("/:id", async (req, res, next) => {
  const serviceFeedback = await services.retrieveProductByIdService(req.params.id);
  return util.sendSuccess(res, serviceFeedback);
});

router.delete(
  "/delete/:productID",
  middlewares.validate(DTO.deleteProductDTO),
  async (req, res, next) => {
    const serviceFeedback = await services.deleteProductService(req.params.productID);
    return util.sendSuccess(res, serviceFeedback);
  },
);

export default router;